#WA-43: Refactor code into service to handle transaction
