+++ src/main/java/com/wildadventure/trip/services/ICommentService.java
@@ -2,7 +2,9 @@ package com.wildadventure.trip.services;
 
 import java.util.List;
 
+import com.wildadventure.trip.models.AddCommentRequest;
 import com.wildadventure.trip.models.Comment;
+import com.wildadventure.trip.models.Trip;
 
 
 public interface ICommentService {
@@ -17,5 +19,5 @@ public interface ICommentService {
 	 * Add a comment associated to a user and a trip
 	 * @param comment
 	 */
-	public Comment addComment(Comment comment);
+	public Comment addComment(AddCommentRequest comment, Trip trip);
 }
